app-2023MT03587/requirement.txt
fastapi
uvicorn
python-dotenv
prometheus_fastapi_instrumentator